monitoring/requirements.txt
requests>=2.25.0
pandas>=1.3.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
gspread>=5.0.0
gspread-dataframe>=3.0.0
google-auth>=2.0.0